+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/ChatEvents/NewChatTitle.kt
@@ -1,7 +1,7 @@
 package dev.inmo.tgbotapi.types.message.ChatEvents
 
-import dev.inmo.tgbotapi.types.message.ChatEvents.abstracts.CommonEvent
+import dev.inmo.tgbotapi.types.message.ChatEvents.abstracts.CommonGroupEvent
 
 data class NewChatTitle(
     val title: String
-): CommonEvent
+): CommonGroupEvent